+++ audiomate/corpus/io/common_voice.py
@@ -103,13 +103,7 @@ class CommonVoiceReader(base.CorpusReader):
             issuer = assets.Speaker(idx, gender=gender, age_group=age)
             corpus.import_issuers(issuer)
             utterance = corpus.new_utterance(idx, idx, issuer.idx)
-
-            utterance.set_label_list(assets.LabelList(idx='word-transcript', labels=[
-                assets.Label(transcription)
-            ]))
-            utterance.set_label_list(assets.LabelList(idx='domain', labels=[
-                assets.Label('speech')
-            ]))
+            utterance.set_label_list(assets.LabelList.create_single(transcription, idx=assets.LL_WORD_TRANSCRIPT))
 
             utt_ids.append(idx)