+++ audiomate/corpus/io/mozilla_deepspeech.py
@@ -5,6 +5,7 @@ import numpy as np
 import scipy
 
 from . import base
+from audiomate.corpus import assets
 from audiomate.utils import textfile
 
 
@@ -20,7 +21,7 @@ class MozillaDeepSpeechWriter(base.CorpusWriter):
         transcription_label_list_idx (str): The transcriptions are used from the label-list with this id.
     """
 
-    def __init__(self, transcription_label_list_idx='default'):
+    def __init__(self, transcription_label_list_idx=assets.LL_WORD_TRANSCRIPT):
         self.transcription_label_list_idx = transcription_label_list_idx
 
     @classmethod