+++ audiomate/corpus/io/tatoeba.py
@@ -170,11 +170,6 @@ class TatoebaReader(base.CorpusReader):
                 corpus.import_issuers(issuer)
 
             utterance = corpus.new_utterance(idx, idx, speaker_idx)
-            utterance.set_label_list(assets.LabelList(idx='word-transcript-raw', labels=[
-                assets.Label(str(transcript))
-            ]))
-            utterance.set_label_list(assets.LabelList(idx='domain', labels=[
-                assets.Label(str('speech'))
-            ]))
+            utterance.set_label_list(assets.LabelList.create_single(transcript, idx=assets.LL_WORD_TRANSCRIPT_RAW))
 
         return corpus