+++ audiomate/corpus/io/voxforge.py
@@ -163,14 +163,11 @@ class VoxforgeReader(base.CorpusReader):
 
                     corpus.new_file(wav_path, idx)
                     utt = corpus.new_utterance(idx, idx, issuer.idx)
-                    utt.set_label_list(assets.LabelList(idx='transcription', labels=[
-                        assets.Label(prompts[basename])
-                    ]))
+                    utt.set_label_list(assets.LabelList.create_single(prompts[basename], idx=assets.LL_WORD_TRANSCRIPT))
 
                     if basename in prompts_orig.keys():
-                        utt.set_label_list(assets.LabelList(idx='transcription_raw', labels=[
-                            assets.Label(prompts_orig[basename])
-                        ]))
+                        raw = assets.LabelList.create_single(prompts_orig[basename], idx=assets.LL_WORD_TRANSCRIPT_RAW)
+                        utt.set_label_list(raw)
 
         return corpus